consolidating codes and files // starting About page
