+++ portfolio/pages/About.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const About = () => {
+
+    return (
+        <div className='w-full'>
+            
+        </div>
+    )
+}
+
+export default About
