+++ portfolio/pages/components/Article.jsx
@@ -18,7 +18,7 @@ const Article = ({title, description, time}) => {
                 <p className="relative z-10 mt-2 text-sm text-zinc-600 dark:text-zinc-400">
                     {description}
                 </p>
-                <div aria-hidden="true" className="relative z-10 mt-4 flex items-center text-sm font-medium text-teal-500">Read article</div>
+                <div aria-hidden="true" className="relative z-10 mt-4 flex items-center text-sm font-medium text-sky-300 underline">Read full blog</div>
             </article>
         </>
     )
