+++ portfolio/pages/index.js
@@ -1,9 +1,6 @@
 import Head from "next/head";
 import Main from "./components/Main";
-import About from "./components/About";
-import Skills from "./components/Skills";
-import Projects from "./Projects";
-import Contact from "./components/Contact";
+import Recent from "./components/recent";
 
 export default function Home() {
   return (
@@ -11,16 +8,16 @@ export default function Home() {
       <Head>
         <title>Tony Bui - Web developper</title>
         <meta name="description" content="Personal website - web developer" />
-        <link rel='icon' href='/PRpic.png' type="image/png" />
+        <link rel="icon" href="/PRpic.png" type="image/png" />
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-        <link href="https://fonts.googleapis.com/css2?family=Sen:wght@400..800&display=swap" rel="stylesheet"></link>
+        <link
+          href="https://fonts.googleapis.com/css2?family=Sen:wght@400..800&display=swap"
+          rel="stylesheet"
+        ></link>
       </Head>
       <Main />
-      <About />
-      <Skills />
-      <Projects />
-      <Contact />
+      <Recent/>
     </div>
   );
 }
